[#144] chore - En detail page 구조 정리
