--- src/pages/DetailPage/DetailPage.module.scss
@@ -1,146 +1,0 @@
-@import 'styles/_palette.scss';
-
-.detail {
-  width: 100%;
-  text-align: center;
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-
-  .product-details-image {
-    max-width: 600px;
-    border-radius: 10px;
-    margin: 30px auto auto;
-    height: auto;
-  }
-
-  .product-details-info {
-    box-sizing: border-box;
-    width: 100%;
-    max-width: 600px;
-    padding: 30px 40px;
-    margin: auto;
-    margin-top: 30px;
-    border-radius: 20px;
-    background-color: $light-gray;
-    text-align: left;
-
-    .product-details-info-description {
-      font-weight: 600;
-      color: $gray;
-      margin-bottom: 10px;
-    }
-
-    .product-details-info-category {
-      font-size: 12px;
-      font-weight: 500;
-      margin-bottom: 25px;
-    }
-
-    .product-details-info-name {
-      font-weight: 600;
-      font-size: 25px;
-      margin-bottom: 32px;
-    }
-
-    .product-details-info-price-info {
-      display: flex;
-      align-items: center;
-      margin-bottom: 10px;
-
-      .product-details-info-price-info-price {
-        margin-left: auto;
-        display: flex;
-        align-items: flex-end;
-        font-size: 25px;
-        font-weight: 700;
-        margin-bottom: 2rem;
-      }
-
-      .product-details-info-price-unit {
-        font-size: 17px;
-        font-weight: normal;
-        margin-left: 0.3rem;
-        margin-right: 0.3rem;
-      }
-    }
-
-    .product-details-funding-info {
-      display: flex;
-      justify-content: space-between;
-      margin-bottom: 20px;
-
-      .product-funding-duedate {
-        font-size: 20px;
-        font-weight: 700;
-      }
-
-      .product-funding-money {
-        font-size: 20px;
-        font-weight: 700;
-      }
-    }
-
-    .product-details-funding {
-      display: flex;
-      justify-content: space-around;
-      align-items: center;
-      background-color: $white;
-
-      .product-funding-percent {
-        margin-right: auto;
-        height: 0.3rem;
-        background-color: $wadiz;
-      }
-    }
-
-    .product-details-info-options {
-      display: flex;
-      justify-content: space-around;
-      align-items: center;
-      padding-top: 30px;
-      margin-top: 10px;
-      border-top: 1px solid rgba(0, 0, 0, 0.1);
-
-      .product-details-info-options-jjim {
-        color: $pink;
-        width: 40px;
-        height: auto;
-        cursor: pointer;
-        animation: heart;
-        animation-duration: 0.3s;
-      }
-
-      .product-details-info-options-funding {
-        color: $wadiz;
-        width: 40px;
-        height: auto;
-        cursor: pointer;
-      }
-
-      @keyframes heart {
-        0% {
-          transform: scale(1);
-        }
-        50% {
-          transform: scale(1.15);
-        }
-        100% {
-          transform: scale(1);
-        }
-      }
-    }
-  }
-
-  .product-more-details-info {
-    box-sizing: border-box;
-    width: 100%;
-    max-width: 600px;
-    padding: 30px 40px;
-    margin: auto;
-    margin-top: 30px;
-    border-radius: 20px;
-    background-color: $light-gray;
-    text-align: left;
-  }
-}
