+++ src/pages/index.js
@@ -1,6 +1,6 @@
 export { default as HomePage } from './HomePage';
 export { default as CategoryPage } from './CategoryPage';
-export { default as DetailPage } from './DetailPage';
+export { default as DetailPage } from './DetailPage/EnDetailPage';
 export { default as MyCartPage } from './MyCartPage';
 export { default as MyHistoryPage } from './MyHistoryPage';
 export { default as MyJjimPage } from './MyJjimPage';
